Only renew certificate if expires within 30 days
This is a quick hack to not hammer the LE API and thus not hit the
weekly limit, if, for instance, your container keeps restarting.

The openssl incantation will only check the certificate expiration time,
and not whether it is valid (ie. checking against the system CA store) -
implementing that is left as an exercise to the reader.

We also make the renew check happen every day, instead of every 60 days,
now that we can afford it.

Signed-off-by: Sergiusz Bazanski <q3k@q3k.org>